split streaming into multiple files
