+++ src/daemon/h2o-common.c
