+++ src/database/engine/dbengine-unittest.c
@@ -108,13 +108,13 @@ static RRDHOST *dbengine_rrdhost_find_or_create(char *name) {
         default_rrd_history_entries,
         RRD_MEMORY_MODE_DBENGINE,
         health_plugin_enabled(),
-        default_rrdpush_enabled,
-        default_rrdpush_destination,
-        default_rrdpush_api_key,
-        default_rrdpush_send_charts_matching,
-        default_rrdpush_enable_replication,
-        default_rrdpush_seconds_to_replicate,
-        default_rrdpush_replication_step,
+        stream_conf_send_enabled,
+        stream_conf_send_destination,
+        stream_conf_send_api_key,
+        stream_conf_send_charts_matching,
+        stream_conf_replication_enabled,
+        stream_conf_replication_period,
+        stream_conf_replication_step,
         NULL,
         0
         );
