+++ src/streaming/stream-capabilities.c
