+++ src/streaming/stream-capabilities.h
