+++ src/streaming/stream-compression/lz4.c
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#include "compression_lz4.h"
+#include "lz4.h"
 
 #ifdef ENABLE_LZ4
 #include "lz4.h"
