+++ src/streaming/stream-compression/lz4.h
