+++ src/web/api/v1/api_v1_info.c
@@ -161,7 +161,7 @@ static int web_client_api_request_v1_info_fill_buffer(RRDHOST *host, BUFFER *wb)
     buffer_json_member_add_uint64(wb, "page-cache-size", default_rrdeng_page_cache_mb);
 #endif // ENABLE_DBENGINE
     buffer_json_member_add_boolean(wb, "web-enabled", web_server_mode != WEB_SERVER_MODE_NONE);
-    buffer_json_member_add_boolean(wb, "stream-enabled", default_rrdpush_enabled);
+    buffer_json_member_add_boolean(wb, "stream-enabled", stream_conf_send_enabled);
 
     buffer_json_member_add_boolean(wb, "stream-compression",
                                    host->sender && host->sender->compressor.initialized);
