+++ src/web/server/h2o/http_server.c
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #include "daemon/common.h"
-#include "streaming/common.h"
+#include "streaming/h2o-common.h"
 #include "http_server.h"
 
 #pragma GCC diagnostic push
